+++ docker-compose.yml
@@ -55,7 +55,8 @@ services:
     ports:
       - "3306:3306"
     volumes:
-      - mysql-data:/var/lib/mysql
+      # - mysql-data:/var/lib/mysql
+      - mysql-data:/var/lib/sora-mysql
       - ./my.cnf:/etc/mysql/my.cnf
     networks:
       - backend
