+++ sora/src/main/java/magicofconch/sora/review/api/dto/res/InquiryMonthRes.java
@@ -1,8 +1,8 @@
-package magicofconch.sora.review.dto.res;
+package magicofconch.sora.review.api.dto.res;
 
 import lombok.Getter;
-import magicOfConch.review.Review;
 import magicOfConch.enums.FeedbackType;
+import magicOfConch.review.Review;
 
 @Getter
 public class InquiryMonthRes {
@@ -10,8 +10,7 @@ public class InquiryMonthRes {
 	private int day;
 	private FeedbackType feedbackType;
 
-
-	public InquiryMonthRes(Review review){
+	public InquiryMonthRes(Review review) {
 		this.day = review.getReviewDate().getDayOfMonth();
 		this.feedbackType = review.getFeedbackType();
 	}
