+++ sora/src/main/java/magicofconch/sora/review/api/dto/res/ReviewRes.java
@@ -1,4 +1,4 @@
-package magicofconch.sora.review.dto.res;
+package magicofconch.sora.review.api.dto.res;
 
 import lombok.Builder;
 import lombok.Getter;
