+++ sora/src/main/java/magicofconch/sora/review/api/dto/res/SubmitRes.java
@@ -1,4 +1,4 @@
-package magicofconch.sora.review.dto.res;
+package magicofconch.sora.review.api.dto.res;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
