+++ sora/src/main/java/magicofconch/sora/user/domain/dto/StreakInfoDto.java
@@ -1,0 +1,5 @@
+package magicofconch.sora.user.domain.dto;
+
+public class StreakInfoDto {
+
+}
